tests: Move negation test to a separate file
